refactor defaultTheme and fix dep cycle
