+++ packages/mui-joy/src/Badge/BadgeProps.ts
@@ -1,8 +1,7 @@
+import * as React from 'react';
 import { ExtendBadgeUnstyledTypeMap } from '@mui/base/BadgeUnstyled';
 import { OverridableStringUnion, OverrideProps } from '@mui/types';
-import * as React from 'react';
-import { SxProps } from '../styles/defaultTheme';
-import { ColorPaletteProp, VariantProp } from '../styles/types';
+import { ColorPaletteProp, VariantProp, SxProps } from '../styles/types';
 
 export type BadgeSlot = 'root' | 'badge';
