+++ packages/mui-joy/src/ListDivider/ListDividerProps.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { OverridableStringUnion, OverrideProps } from '@mui/types';
-import { SxProps } from '../styles/defaultTheme';
+import { SxProps } from '../styles/types';
 import { ListDividerClasses } from './listDividerClasses';
 
 export type ListDividerSlot = 'root';
