+++ packages/mui-joy/src/ListItem/ListItemProps.ts
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { OverridableStringUnion, OverrideProps } from '@mui/types';
-import { SxProps } from '../styles/defaultTheme';
-import { ColorPaletteProp, VariantProp } from '../styles/types';
+import { ColorPaletteProp, VariantProp, SxProps } from '../styles/types';
 import { ListItemClasses } from './listItemClasses';
 
 export type ListItemSlot = 'root' | 'startAction' | 'endAction';
