+++ packages/mui-joy/src/styles/styled.ts
@@ -1,7 +1,8 @@
 import { createStyled } from '@mui/system';
-import defaultTheme, { JoyTheme } from './defaultTheme';
+import { Theme } from './types';
+import defaultTheme from './defaultTheme';
 import styleFunctionSx from './styleFunctionSx';
 
-const styled = createStyled<JoyTheme>({ defaultTheme, styleFunctionSx });
+const styled = createStyled<Theme>({ defaultTheme, styleFunctionSx });
 
 export default styled;
