+++ packages/mui-joy/src/styles/types/index.ts
@@ -5,3 +5,4 @@ export * from './radius';
 export * from './shadow';
 export * from './variants';
 export * from './typography';
+export * from './theme';
