+++ packages/mui-joy/src/styles/types/theme.ts
@@ -1,0 +1,76 @@
+import {
+  Breakpoints,
+  Spacing,
+  SxProps as SystemSxProps,
+  SystemProps as SystemSystemProps,
+} from '@mui/system';
+import { DefaultColorScheme, ExtendedColorScheme } from './colorScheme';
+import { ColorSystem } from './colorSystem';
+import { Focus } from './focus';
+import { Shadow } from './shadow';
+import { Radius } from './radius';
+import {
+  FontFamily,
+  FontSize,
+  FontWeight,
+  LineHeight,
+  LetterSpacing,
+  TypographySystem,
+} from './typography';
+import { Variants } from './variants';
+
+type Split<T, K extends keyof T = keyof T> = K extends string | number
+  ? { [k in K]: Exclude<T[K], undefined> }
+  : never;
+
+type ConcatDeep<T> = T extends Record<string | number, infer V>
+  ? keyof T extends string | number
+    ? V extends string | number
+      ? keyof T
+      : keyof V extends string | number
+      ? `${keyof T}-${ConcatDeep<Split<V>>}`
+      : never
+    : never
+  : never;
+
+type NormalizeVars<T> = ConcatDeep<Split<T>>;
+
+export interface RuntimeColorSystem extends Omit<ColorSystem, 'palette'> {
+  palette: ColorSystem['palette'] & {
+    mode: 'light' | 'dark';
+    colorScheme: DefaultColorScheme | ExtendedColorScheme;
+  };
+}
+
+export interface ThemeScales {
+  radius: Radius;
+  shadow: Shadow;
+  fontFamily: FontFamily;
+  fontSize: FontSize;
+  fontWeight: FontWeight;
+  lineHeight: LineHeight;
+  letterSpacing: LetterSpacing;
+}
+
+export interface ThemeVars extends ThemeScales, ColorSystem {}
+
+export type ThemeCSSVar = NormalizeVars<ThemeVars>;
+
+export interface Theme extends ThemeScales, RuntimeColorSystem {
+  colorSchemes: Record<DefaultColorScheme | ExtendedColorScheme, ColorSystem>;
+  focus: Focus;
+  typography: TypographySystem;
+  variants: Variants;
+  spacing: Spacing;
+  breakpoints: Breakpoints;
+  prefix: string;
+  vars: ThemeVars;
+  getCssVar: <CustomVar extends string = never>(
+    field: ThemeCSSVar | CustomVar,
+    ...vars: (ThemeCSSVar | CustomVar)[]
+  ) => string;
+}
+
+export type SxProps = SystemSxProps<Theme>;
+
+export type SystemProps = SystemSystemProps<Theme>;
